// src/types/chat.ts
export type ChatState = {
    uid: string,
    menu: string,
    gappi: string,
    tomare: [],
    tomareId: string,
    am_pm: string
    make: boolean
    nail: boolean
    este: boolean
    sonota: string
    yoyakuId: string
    message: string
    timestamp: string
    yoyakuUid: string
    yoyakuIcon: string
    yoyakuMenu: string
}